Version 4.6.6
Status Module:
Set your own Status or select one of the default, you can customize the Default Status in the Config System.

Language System:
The Language System is now on the Way, I want to have good translations, so you can help to translate on a website. (Link will come with full release)

Mainteance Feature:
This Feature is now finished and work now how I want that it work.

From now on every new Feature become a Version Tag and since when it is in the Plugin.