Full Changelog Version 4.6.6:
Updated for Version 1.19

Status Module → Set your own Status or selected of one of the default Status which can be set by the Server Owner.

Maintenance System is now fixed and finished and should work.

Plugin has now an own Language System with the Languages German and English, with your help there will come more.
Customize 99% of the Plugin Messages

Removed the AFKKickOnDamage and Custom Crafting Recipes (CCR)
CCR will come back in the Version 4.6.7 with a better Version and an own GUI

Commands like Heal or GM (Gamemode) will also be usable on other Players when Version 4.6.7 comes out.