« Version 4.6.8.A1 Changelog »
WARNING: THIS IS ONLY AN ALPHA VERSION. USE THIS ONLY FOR TESTING. THERE IS NO SUPPORT!

Download: NO DOWNLOAD AVAILABLE - BUILD YOURSELF VERSION

» 1.19.4 Support
The Server Version 1.19.4 is now official Supported with this Version.

» Discord Module Rework
Added in Discord the /mcinfo which shows Infos about the Minecraft Server

«================================»

Please report any bugs on the Discord (https://yvnetwork.de/yvtils-smp/support) or with Private Message to @WolfiiYV#3204.