Version 4.6.8.1 ALPHA Changelog
» Sit
    Removed the ability to sit down when clicking on a stair.

» CCR
    Added support to craft multiple items at once.
    Added a new recipe; more information will follow.

» Bug Fixes

«================================»

» REWORK ANNOUNCEMENT
Today, I will announce something big for this plugin! I will release Version 4.6.8.1 in full, and after that, I will start to rework the entire plugin from the beginning. I've noticed that I've lost sight of everything here, so I want to restructure, fix bugs, add long-planned features, and more.

It will probably take a while until I am finished, as I have a lot to do. But when I am finished, everything will be better—more customizability, and probably more of everything! (We will see.)

To see what is finished, it would probably be best to look on GitHub in the TODO file. I will mark everything I am done with an "x" in front of it.

Bug fixes will still be released, I think. Also, if a new minecraft version needs an update, I will try to work on it with the old code or the partially reworked code.

See you!

«================================»

» YVtils Adventures
This plugin (mentioned in an older changelog) is now in the testing phase. There are still missing translations and features, but I will pause the coding progress for now and work on this plugin.

«================================»

Please report any bugs on Discord (https://yvnetwork.de/yvtils-smp/support) or through a private message to @WolfiiYV#3204.